Rename statistics.py to metrics.py
